Add config.production.json in backup, not in restore, restore settings
Ghost docker image has config.production.json installed alongside Ghost code.
When mounted into another container that file is not visible.
Also, it is common to bind mount the config file from the host in which
case it cannot be removed either (Device or resource busy)

Therefore, it's still preferable to include the config in the archive
but there shouldn't be attempt to remove it in restore.

There is a "settings" directory in /var/lib/ghost/content/ that the
default restore behaviour didn't know about, so added it for removal.